mart_transport/requirements.txt
duckdb==0.9.2
mysql-connector-python==8.2.0
pandas==2.1.4
python-dotenv==1.0.0
